% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_guides.R
\name{add_guides}
\alias{add_guides}
\title{Concatenate full length guides and short read data.}
\usage{
add_guides(
  dir_out = ".",
  filename_out = "seqs_and_guides.fasta",
  guides_file_path,
  asvs_file_path
)
}
\arguments{
\item{dir_out}{The output directory where to write the concatenated ASV + guides file.}

\item{filename_out}{Output filename to write the concatenated ASV + guides file.}

\item{guides_file_path}{File path to the full length guide file.}

\item{asvs_file_path}{File path to ASV file (FASTA formatted).}
}
\value{
The path to the ouput concatenated ASV + guides file.
}
\description{
Concatenate full length guides and short read data.
}
