% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ex_ps.R
\docType{data}
\name{ex_ps}
\alias{ex_ps}
\title{Example phyloseq object with bad ASV sequence}
\format{
\subsection{\code{ex_ps}}{

A phyloseq object generated from the DADA2 Pipeline tutorial data 1.16. This data has an example "bad ASV" sequence added for tutorial puposes.
\describe{
\item{country}{USA}
\item{year}{2025}
...
}
}
}
\source{
\url{https://benjjneb.github.io/dada2/tutorial.html}
}
\usage{
ex_ps
}
\description{
The phyloseq data
Report ...
}
\keyword{datasets}
