% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guides_data.R
\docType{data}
\name{guides_data}
\alias{guides_data}
\title{Example phyloseq object}
\format{
\subsection{\code{guides_data}}{

A phyloseq object generated from the DADA2 Pipeline tutorial data 1.16. This data has an example "bad ASV" sequence added for tutorial puposes.
\describe{
\item{country}{USA}
\item{year}{2025}
...
}
}
}
\source{
\url{https://benjjneb.github.io/dada2/tutorial.html}
}
\usage{
guides_data
}
\description{
The phyloseq data
Report ...
}
\keyword{datasets}
