% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_fasttree_available.R
\name{is_fasttree_available}
\alias{is_fasttree_available}
\title{Check if fasttree is available on the system}
\usage{
is_fasttree_available()
}
\value{
Returns the number of characters of path for FastTree
}
\description{
Check if fasttree is available on the system
}
