% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_mothur_available.R
\name{is_mothur_available}
\alias{is_mothur_available}
\title{Check if mothur is available on the system}
\usage{
is_mothur_available()
}
\value{
Returns the number of characters of path for mothur
}
\description{
Check if mothur is available on the system
}
