% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_fasttree.R
\name{make_fasttree}
\alias{make_fasttree}
\title{Make a fasttree from an alignment file}
\usage{
make_fasttree(
  path_alignment,
  fasttree_flags = "-nt -gtr -gamma",
  output_file_name,
  output_dir
)
}
\arguments{
\item{path_alignment}{Path to the alignment file}

\item{fasttree_flags}{Flags to pass to fasttree}

\item{output_file_name}{The output file name}

\item{output_dir}{The output directory.}
}
\value{
Returns a path to fasttree output if fasttree is successful,
otherwise returns NULL
}
\description{
Make a fasttree from an alignment file
}
