% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_taxa_with_short_names.R
\name{rename_taxa_with_short_names}
\alias{rename_taxa_with_short_names}
\title{Rename taxa in a phyloseq object using a fasta key}
\usage{
rename_taxa_with_short_names(physeq, fasta_df)
}
\arguments{
\item{physeq}{A phyloseq object whose taxa_names are long ASV sequences.}

\item{fasta_df}{A data frame with two columns: "seq.name" containing human-readable
identifiers (e.g., "seq1", "seq2") and "seq.text" containing the corresponding ASV sequences.
This can be generated by reading in a fasta file using tools like phylotools' read.fasta function.}
}
\value{
Returns a phyloseq object where the taxa names (ASV sequences) are replaced with the
corresponding short names from the fasta key.
If any sequences in the phyloseq object are not found in the key, they will be left unchanged
and a warning will be issued.
}
\description{
Rename taxa in a phyloseq object using a fasta key
}
