% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence_key_to_fasta.R
\name{sequence_key_to_fasta}
\alias{sequence_key_to_fasta}
\title{Write a fasta file from a sequence key}
\usage{
sequence_key_to_fasta(file, seqkey)
}
\arguments{
\item{file}{Path to file to write out.}

\item{seqkey}{A sequence key}
}
\value{
Returns the file path to the sequence key
}
\description{
Write a fasta file from a sequence key
}
