% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_guides.R
\name{add_guides}
\alias{add_guides}
\title{Concatenate guides and fasta}
\usage{
add_guides(
  dir_out = ".",
  filename_out = "seqs_and_guides.fasta",
  guides_file_path,
  asvs_file_path
)
}
\arguments{
\item{dir_out}{The output directory where to write the seqs + guides file.}

\item{filename_out}{the name of the output file of the seqs + guides file}

\item{guides_file_path}{The file path to the guide file}

\item{asvs_file_path}{The file path to asv fasta file}
}
\value{
The path to the written fasta + guides file
}
\description{
Concatenate guides and fasta
}
