% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align_seqs_mothur.R
\name{align_seqs_mothur}
\alias{align_seqs_mothur}
\title{Wrapper for mothur's align.seqs() function.
Convience R script to use the mothur align.seqs command. More information can be
found at the mothur website: https://mothur.org/wiki/align.seqs/.
Defaults are set to those defaults provided by the documentation at
mothur.org}
\usage{
align_seqs_mothur(
  candidate,
  output.directory,
  template,
  search = "kmer",
  ksize = 8,
  align = "needleman",
  match = 1,
  mismatch = -1,
  gapopen = -2,
  gapextend = -1,
  flip = TRUE,
  threshold = 0.5,
  processors = 1
)
}
\arguments{
\item{candidate}{Path to the fasta file that you are wanting to align (guides + ASV concatenated file)}

\item{output.directory}{Output directory for which to write output files.}

\item{template}{The path to the template alignment file for which NAST alignment algorithm will align to.}

\item{search}{Set search to "kmer" (Default) or "suffix" for kmer or suffix tree searching. The kmer searching is recommended on mothur.org with cited reasons of being faster and best.}

\item{ksize}{The size of the kmer. mothur.org recommends the kmer size that is empirically determined by the user to be the fastest for alignment.}

\item{align}{The alignment method used. Options include "needleman" (Default) and "gotoh" algorithm. The needleman prioritizes the same amount for opening and extending a gap. The gotoh algorithm charges differently for opening (larger penalty) and extending a gap. mothur.org suggests the needleman algorithm as default.}

\item{match}{Reward for a match (Default +1). mothur.org has set these defaults to be best at producing 16S rRNA gene sequence alignments.}

\item{mismatch}{The penalty for a mismatch (Default -1).}

\item{gapopen}{The penalty for opening a gap (Default is -2).}

\item{gapextend}{The  penalty for extending a gap (Default -1).}

\item{flip}{Should the sequence be flipped to align to see if the reverse complement aligns better? (Default is TRUE).}

\item{threshold}{The threshold (Default 0.5) that is used to determine if a sequence should be flipped for a better alignment.}

\item{processors}{The number of cores to use to do the alignment. Default 1.}
}
\value{
A path to the alignment file produced by mothurs align.seqs function.
}
\description{
Wrapper for mothur's align.seqs() function.
Convience R script to use the mothur align.seqs command. More information can be
found at the mothur website: https://mothur.org/wiki/align.seqs/.
Defaults are set to those defaults provided by the documentation at
mothur.org
}
