% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ex_ps.R
\docType{data}
\name{ex_ps}
\alias{ex_ps}
\title{Example phyloseq object with a single bad ASV sequence which will produce a poor quality alignment for tutorial purposes.}
\format{
\subsection{\code{ex_ps}}{

A phyloseq object generated from the DADA2 Pipeline tutorial data 1.16. This data has an example "bad ASV" sequence added for tutorial purposes to simulate what to do for low quality alignments.
\describe{
\item{country}{USA}
\item{year}{2025}
...
}
}
}
\source{
\url{https://benjjneb.github.io/dada2/tutorial.html}
}
\usage{
ex_ps
}
\description{
Example phyloseq object with a single bad ASV sequence which will produce a poor quality alignment for tutorial purposes.
}
\keyword{datasets}
